moved TypeHelper methods to TypeExtensions
